# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a projectc-custodial-wallet.",
        "title": "Swagger projectc-custodial-wallet",
        "contact": {
            "name": "guyuxiang",
            "url": "https://guyuxiang.github.io",
            "email": "gu_yuxiang@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/inner/wallet/callback/rollback": {
            "post": {
                "description": "Receive rollback callback pushed by connector and mark related wallet transactions as failed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "ReceiveRollbackCallback",
                "parameters": [
                    {
                        "description": "Connector transaction rollback callback request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ConnectorTxRollbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/inner/wallet/callback/tx": {
            "post": {
                "description": "Receive transaction callback pushed by connector and update wallet transactions.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "ReceiveTxCallback",
                "parameters": [
                    {
                        "description": "Connector transaction callback request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ConnectorTxCallbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
        "/wallet/create": {
            "post": {
                "description": "Create a master walletNo and automatically create addresses for all configured networks.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "CreateWallet",
                "parameters": [
                    {
                        "description": "Create wallet request, no parameters required",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.WalletCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/wallet/info/query": {
            "post": {
                "description": "Query token balances of a wallet. When network is specified, return balances under that network; when network is empty, aggregate balances across all supported networks under the master walletNo.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "QueryWalletInfo",
                "parameters": [
                    {
                        "description": "Wallet info query request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.WalletInfoQueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/wallet/transaction/history/query": {
            "post": {
                "description": "Query wallet transaction history by cursor and time range.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "QueryHistory",
                "parameters": [
                    {
                        "description": "Transaction history query request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TransactionHistoryQueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/wallet/transaction/query": {
            "post": {
                "description": "Query a single transaction by transactionNo.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "QueryTransaction",
                "parameters": [
                    {
                        "description": "Transaction query request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TransactionQueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/wallet/transfer/out": {
            "post": {
                "description": "Submit a wallet transfer-out request by tokenSymbol for the specified network.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "TransferOut",
                "parameters": [
                    {
                        "description": "Transfer out request, tokenSymbol is used instead of tokenAddress",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TransferOutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/wallet/transfer/out/query": {
            "post": {
                "description": "Query transferable assets of a wallet under a specific network, including native token and configured network tokens.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "QueryTransferOutAssets",
                "parameters": [
                    {
                        "description": "Transfer out capability query request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.TransferOutQueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ConnectorChainEvent": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "models.ConnectorChainTx": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "blockNumber": {
                    "type": "integer"
                },
                "code": {
                    "type": "string"
                },
                "fee": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "networkCode": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "models.ConnectorTxCallbackRequest": {
            "type": "object",
            "properties": {
                "tx": {
                    "$ref": "#/definitions/models.ConnectorChainTx"
                },
                "txEvents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ConnectorChainEvent"
                    }
                }
            }
        },
        "models.ConnectorTxRollbackRequest": {
            "type": "object",
            "properties": {
                "networkCode": {
                    "type": "string"
                },
                "txCode": {
                    "type": "string"
                }
            }
        },
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.TransactionHistoryQueryRequest": {
            "type": "object",
            "required": [
                "pageSize",
                "walletNo"
            ],
            "properties": {
                "cursor": {
                    "type": "integer"
                },
                "direction": {
                    "type": "string"
                },
                "endTime": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "startTime": {
                    "type": "integer"
                },
                "walletNo": {
                    "type": "string"
                }
            }
        },
        "models.TransactionQueryRequest": {
            "type": "object",
            "required": [
                "transactionNo"
            ],
            "properties": {
                "transactionNo": {
                    "type": "string"
                }
            }
        },
        "models.TransferOutQueryRequest": {
            "type": "object",
            "required": [
                "walletNo"
            ],
            "properties": {
                "network": {
                    "type": "string"
                },
                "walletNo": {
                    "type": "string"
                }
            }
        },
        "models.TransferOutRequest": {
            "type": "object",
            "required": [
                "amount",
                "network",
                "requestNo",
                "toAddress",
                "walletNo"
            ],
            "properties": {
                "amount": {
                    "type": "string"
                },
                "network": {
                    "type": "string"
                },
                "requestNo": {
                    "type": "string"
                },
                "toAddress": {
                    "type": "string"
                },
                "tokenSymbol": {
                    "type": "string"
                },
                "walletNo": {
                    "type": "string"
                }
            }
        },
        "models.WalletCreateRequest": {
            "type": "object"
        },
        "models.WalletInfoQueryRequest": {
            "type": "object",
            "required": [
                "walletNo"
            ],
            "properties": {
                "network": {
                    "type": "string"
                },
                "walletNo": {
                    "type": "string"
                }
            }
        }
    }
}